requirements.txt
openpyxl
pillow
pytesseract
tesseract